Fixes for edges and pins
The B.CrtYd-Layer is now Edge.Cuts including two falsely layered
Eco2.User graphic segments. The board is now perfectly shown in 3D
viewer and on does not have a problem in the PCB layout to find the
correct pcb boundaries (this layout is now not panalizeable anymore).

The paste and mask layer for the B-pins were falsely on the F-layer.